+++ TProject/TypeDAO/StreetDAO.cs
@@ -1,4 +1,5 @@
-﻿using System.Data.SQLite;
+﻿using System.Collections.Generic;
+using System.Data.SQLite;
 using System.Linq;
 
 namespace TProject.TypeDAO
@@ -29,8 +30,24 @@ namespace TProject.TypeDAO
 
         public bool Delete(string name)
         {
+            bool flag = true;
+
             try
             {
+                List<List<object>> list = GetAll("Edge");
+                list.ForEach((e) =>
+                {
+                    if (e[5].ToString().Equals(name))
+                    {
+                        flag = false;
+                    }
+                });
+
+                if (!flag)
+                {
+                    return false;
+                }
+
                 new SQLiteCommand(string.Format("DELETE from Street where Name = '{0}'", name.ToLower()), DAO.GetConnection()).ExecuteNonQuery();
 
                 Graph.Edge.StreetList.RemoveAll(l => l.ElementAt(0).ToString().ToLower().Equals(name.ToLower()));
@@ -45,8 +62,24 @@ namespace TProject.TypeDAO
 
         public bool Update(object obj, string ID)
         {
+            bool flag = true;
+
             try
             {
+                List<List<object>> list = GetAll("Edge");
+                list.ForEach((e) =>
+                {
+                    if (e[5].ToString().Equals(ID))
+                    {
+                        flag = false;
+                    }
+                });
+
+                if (!flag)
+                {
+                    return false;
+                }
+
                 new SQLiteCommand(string.Format("UPDATE Street SET [Name] = '{0}' where Name = '{1}'", obj.ToString().ToLower(), ID.ToLower()), DAO.GetConnection()).ExecuteNonQuery();
 
                 var updatedStreet = Graph.Edge.StreetList.First(l => l.ElementAt(0).ToString().ToLower().Equals(ID.ToLower()));
